+++ src/vid.hpp
@@ -1,12 +1,11 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 namespace splonks {
 
 struct VID {
-    std::size_t id = 0;
+    std::uint32_t id = 0;
     std::uint32_t version = 0;
 };
